+++ OUDS/Core/ThemesContract/Sources/TokenatorConstants.swift
@@ -84,14 +84,14 @@ public enum OUDSVersions {
     public static let componentPinCodeInputVersion = "1.3.0"
     /// Version of the Figma specifications for the component quantity input (1.3.1)
     public static let componentQuantityInputVersion = "1.3.1"
-    /// Version of the Figma specifications for the component quantity input (1.0.0)
-    public static let componentStaticListItemVersion = "1.0.0"
-    /// Version of the Figma specifications for the component quantity input (1.0.0)
-    public static let componentStaticCardItemVersion = "1.0.0"
-    /// Version of the Figma specifications for the component quantity input (1.0.0)
-    public static let componentNavigationListItemVersion = "1.0.0"
-    /// Version of the Figma specifications for the component quantity input (1.0.0)
-    public static let componentNavigationCardItemVersion = "1.0.0"
+    /// Version of the Figma specifications for the component quantity input (1.2.0)
+    public static let componentStaticListItemVersion = "1.2.0"
+    /// Version of the Figma specifications for the component quantity input (1.2.0)
+    public static let componentStaticCardItemVersion = "1.2.0"
+    /// Version of the Figma specifications for the component quantity input (1.2.0)
+    public static let componentNavigationListItemVersion = "1.2.0"
+    /// Version of the Figma specifications for the component quantity input (1.2.0)
+    public static let componentNavigationCardItemVersion = "1.2.0"
     /// Version of the Figma specifications for the component radio button (1.4.0)
     public static let componentRadioButtonVersion = "1.4.0"
     /// Version of the Figma specifications for the component select input (1.3.1)